# Avoid allocs improving ~15% performance and using ~97% less memory