--- a/bandersnatch/multiexp.go
+++ b/bandersnatch/multiexp.go
@@ -392,67 +392,9 @@ func msmProcessChunkPointAffine(chunk uint64,
 	c uint64,
 	points []PointAffine,
 	scalars []fr.Element) {
-
-	mask := uint64((1 << c) - 1) // low c bits are 1
-	msbWindow := uint64(1 << (c - 1))
-
-	for i := 0; i < len(buckets); i++ {
-		buckets[i].Identity()
-	}
-
-	jc := uint64(chunk * c)
-	s := selector{}
-	s.index = jc / 64
-	s.shift = jc - (s.index * 64)
-	s.mask = mask << s.shift
-	s.multiWordSelect = (64%c) != 0 && s.shift > (64-c) && s.index < (fr.Limbs-1)
-	if s.multiWordSelect {
-		nbBitsHigh := s.shift - uint64(64-c)
-		s.maskHigh = (1 << nbBitsHigh) - 1
-		s.shiftHigh = (c - nbBitsHigh)
-	}
-
-	// for each scalars, get the digit corresponding to the chunk we're processing.
-	for i := 0; i < len(scalars); i++ {
-		bits := (scalars[i][s.index] & s.mask) >> s.shift
-		if s.multiWordSelect {
-			bits += (scalars[i][s.index+1] & s.maskHigh) << s.shiftHigh
-		}
-
-		if bits == 0 {
-			continue
-		}
-
-		// if msbWindow bit is set, we need to substract
-		if bits&msbWindow == 0 {
-			// add
-			var pProj PointProj
-			pProj.FromAffine(&points[i])
-			buckets[bits-1].Add(&pProj, &buckets[bits-1])
-		} else {
-			// sub
-			var pProj PointProj
-			pProj.FromAffine(&points[i])
-			pProj.Neg(&pProj)
-			buckets[bits & ^msbWindow].Add(&buckets[bits & ^msbWindow], &pProj)
-		}
-	}
-
-	// reduce buckets into total
-	// total =  bucket[0] + 2*bucket[1] + 3*bucket[2] ... + n*bucket[n-1]
-
-	var runningSum, total PointProj
-	runningSum.Identity()
-	total.Identity()
-	for k := len(buckets) - 1; k >= 0; k-- {
-
-		runningSum.Add(&runningSum, &buckets[k])
-
-		total.Add(&total, &runningSum)
-	}
-
+	var total PointProj
+	msmProcessChunkPointAffineDMA(chunk, &total, buckets, c, points, scalars)
 	chRes <- total
-
 }
 
 func msmProcessChunkPointAffineDMA(chunk uint64,