--- a/bandersnatch/precomp_multiexp.go
+++ b/bandersnatch/precomp_multiexp.go
@@ -3,10 +3,8 @@ package bandersnatch
 import (
 	"encoding/binary"
 	"io"
-	"runtime"
 
 	"github.com/crate-crypto/go-ipa/bandersnatch/fr"
-	"github.com/crate-crypto/go-ipa/bandersnatch/parallel"
 )
 
 type PrecomputeLagrange struct {
@@ -83,42 +81,28 @@ func DeserializePrecomputedLagrange(reader io.Reader) (*PrecomputeLagrange, erro
 }
 
 func (p *PrecomputeLagrange) Commit(evaluations []fr.Element) *PointAffine {
-	nbTasks := runtime.NumCPU()
-	chPartialResults := make(chan PointProj, nbTasks)
-	parallel.Execute(len(evaluations), func(start, end int) {
-		var partial_result PointProj
-		partial_result.Identity()
-
-		for i := start; i < end; i++ {
-			scalar := &evaluations[i]
-
-			if scalar.IsZero() {
-				continue
-			}
-
-			table := p.inner[i]
-			scalar_bytes_le := scalar.BytesLE()
-
-			for row, byte := range scalar_bytes_le {
-				var tp = table.point(row, byte)
-				var tp_proj PointProj
-				tp_proj.FromAffine(tp)
-				partial_result.Add(&partial_result, &tp_proj)
-			}
-		}
-
-		chPartialResults <- partial_result
-
-	}, nbTasks)
-
-	// Aggregate Parallel Results
-
-	close(chPartialResults)
 	var result PointProj
 	result.Identity()
-	for partial := range chPartialResults {
-		result.Add(&result, &partial)
+
+	for i := 0; i < len(evaluations); i++ {
+		scalar := &evaluations[i]
+
+		if scalar.IsZero() {
+			continue
+		}
+
+		table := p.inner[i]
+		scalar_bytes_le := scalar.BytesLE()
+
+		for row, byte := range scalar_bytes_le {
+			var tp = table.point(row, byte)
+			var tp_proj PointProj
+			tp_proj.FromAffine(tp)
+			result.Add(&result, &tp_proj)
+		}
 	}
+
+	// Aggregate Parallel Results
 
 	var res_affine PointAffine
 	res_affine.FromProj(&result)