@jsign 2022-05-16
> These are amazing results! Thank you very much, I'll check out the PR now :)

You're welcome!

> I should also note that https://github.com/crate-crypto/go-ipa/pull/19 might regress some of the changes made here. Instead of PointAffine, we use a wrapper around PointProj with special methods, so MultiExp etc now also use that wrapper

Sounds good. Maybe after that gets merged I can take some other pass with optimization-eyes to see if there's something extra that we can gain.