@kevaundray 2022-05-16
Alright, I'll merge this first and then resolve conflicts on that branch, aiming to preserve the optimisations you have done